Add ability to add new books
This commit adds the function for the user to add new books
to the library by using a form and dialog. The current plan is
for the table to be refreshed everytime a new book is added but a way
to either clear the new branches or only append the final object in
the array.